Upgraded to upstream version 2013-06-27

+++ hdd/HDD_RemoveDir.c
@@ -4,12 +4,23 @@
 
 void HDD_RemoveDir(char *DirPath, bool Recursive)
 {
+  #ifdef DEBUG_FIREBIRDLIB
+    CallTraceEnter("HDD_RemoveDir");
+  #endif
+
   char                  cmd[512];
 
-  if(Recursive)
-    TAP_SPrint(cmd, "rm -rf %s/%s", TAPFSROOT, DirPath);
-  else
-    TAP_SPrint(cmd, "rmdir %s/%s", TAPFSROOT, DirPath);
+  if(DirPath && *DirPath)
+  {
+    if(Recursive)
+      TAP_SPrint(cmd, "rm -rf %s/%s", TAPFSROOT, DirPath);
+    else
+      TAP_SPrint(cmd, "rmdir %s/%s", TAPFSROOT, DirPath);
 
-  system(cmd);
+    system(cmd);
+  }
+
+  #ifdef DEBUG_FIREBIRDLIB
+    CallTraceExit(NULL);
+  #endif
 }
